Mirror the npm adapter folder layout for PyPI
Split server/lib/adapters/pypi/index.ts (866 lines) to match the npm
adapter's structure (types, manifest, findings, acquire, broker), with
index.ts as the orchestrator + barrel re-exporting the same public
surface:

- types (manifest/artifact types, constants, regexes)
- manifest (parsing, name normalization + validation)
- findings (pyPiReleaseFindings + summarizePyPiArtifact)
- acquire (staged/baseline acquisition, remote fetch, artifact prep)

broker imports its types from ./types. firstMatchingLine imports from
text-utils. Pure refactor, no behavior change.
